refactor(folders): move ensure_folder_hierarchy into folder_service
The derived index needs it too, and importing it from the agent middleware
closed an import cycle. kb_persistence is deleted at the git-native cut, so the
helper had to outlive it either way.
